BUILD-4326 Authenticate with artifactory when downloading O#
